pyproject.toml
[project]
name = "llamaapp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">= 3.12"
dependencies = [
    "chromadb>=1.5.1",
    "crawl4ai>=0.8.0",
    "dotenv-python>=0.0.1",
    "google-generativeai>=0.5.4",
    "llama-index>=0.14.15",
    "llama-index-core>=0.14.15",
    "llama-index-embeddings-gemini>=0.4.2",
    "llama-index-embeddings-google>=0.4.2",
    "llama-index-embeddings-google-genai>=0.3.2",
    "llama-index-embeddings-mistralai>=0.4.1",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-gemini>=0.6.2",
    "llama-index-llms-google-genai>=0.8.7",
    "llama-index-llms-mistralai>=0.9.0",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-readers-web>=0.5.6",
    "llama-index-vector-stores-chroma>=0.5.5",
    "llama-index-vector-stores-pinecone>=0.7.1",
    "mistralai>=1.9.11",
    "pinecone>=7.0.0,<8.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
